code/hejiaqin/phone/src/androidTest/java/com/chinamobile/hejiaqin/business/ui/more/MoreSearchAndSendContactActivityTest.java
package com.chinamobile.hejiaqin.business.ui.more;

import android.content.Intent;
import android.test.ActivityUnitTestCase;
import android.view.View;
import android.widget.EditText;
import android.widget.ListView;

import com.chinamobile.hejiaqin.R;
import com.chinamobile.hejiaqin.business.ui.contact.ContactSearchActivity;

import java.io.Serializable;
import java.util.HashSet;
import java.util.Set;

/**
 * Created by Administrator on 2017/4/24 0024.
 */
public class MoreSearchAndSendContactActivityTest extends
        ActivityUnitTestCase<MoreSearchAndSendContactActivity> {
    private MoreSearchAndSendContactActivity mActivity;

    private EditText searchInput;
    private View searchDelete;
    private View searchCancel;
    private ListView contactsListView;

    public MoreSearchAndSendContactActivityTest() {
        super(MoreSearchAndSendContactActivity.class);
    }

    @Override
    protected void setUp() throws Exception {
        super.setUp();

        Intent intent = new Intent(getInstrumentation().getTargetContext(),
                MoreSearchAndSendContactActivity.class);
        intent.putExtra(ContactSearchActivity.Constant.INTENT_DATA_CONTACT_TYPE,
                ContactSearchActivity.Constant.CONTACT_TYPE_APP);
        Set<String> set = new HashSet<>();
        set.add("aaa");
        intent.putExtra("selected", (Serializable) set);
        startActivity(intent, null, null);

        mActivity = getActivity();
        assertNotNull(mActivity);

        searchInput = (EditText) mActivity.findViewById(R.id.search_input);
        searchDelete = mActivity.findViewById(R.id.search_del);
        searchCancel = mActivity.findViewById(R.id.search_cancel);
        contactsListView = (ListView) mActivity.findViewById(R.id.list);
    }

    public void testPreconditons() {
        assertNotNull(searchInput);
        assertNotNull(searchDelete);
        assertNotNull(searchCancel);
        assertNotNull(contactsListView);
    }
}
